refactor(auth): move JwtPayload to module-core with backward-compatible typealias
isExpired() dropped (dead code — zero callers). Existing module-infra
imports resolve via typealias.
